Conflicts:
	src/open/core/geometry/line.h
	src/open/core/geometry/polygonPointIterator.h
	src/open/core/patterndetection/chessBoardDetector.h
	src/open/core/xml/generated/checkerboardDetectionParameters.cpp
	src/open/core/xml/generated/generated.pri
	src/open/core/xml/patternDetector.xml
	src/open/utils/capture/photostationCaptureDialog.cpp
	src/open/utils/corestructs/coreWidgets/bitcodeBoardParamsBaseControlWidget.ui
	src/open/utils/corestructs/coreWidgets/coreWidgets.pri
	src/open/utils/filters/ui/filterWidgets.pri
	src/open/utils/framesources/directShow/directShowCapture.cpp
	src/open/utils/framesources/directShow/directShowCapture.h
	src/open/utils/utils.pro
	src/open/wrappers/directShow/lib64/capdll.dll
	src/open/wrappers/directShow/lib64/capdll_vc12.lib
	src/restricted/topcon/topcon_restricted.pro
	src/restricted/topcon/utils/topcon_utils_restricted.pri
	src/restricted/topcon/utils/topcon_utils_restricted.pro

+++ test/softrender/main_softrender.cpp
@@ -38,7 +38,7 @@ int main(int argc, const char **argv)
     }
     if (!mesh.hasColor) {
        mesh.switchColor();
-       for (int i = 0;i < mesh.facesColor.size();i++)
+       for (size_t i = 0;i < mesh.facesColor.size();i++)
        {
            mesh.facesColor[i] = RGBColor::rainbow1((double)i / mesh.facesColor.size());
        }
